backend/requirements-web.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.4.2
celery==5.4.0
billiard==4.2.1
psycopg2-binary==2.9.9
pydantic==2.11.0
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
redis==5.0.7
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
vine==5.1.0
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
mcp[cli]==1.13.1
email-validator==2.2.0
pytest==7.4.4
pytest-asyncio==0.23.8
fakeredis==2.23.3

